Update the text of a quiz question's option
This commit pertains to a slight modification in the text of a quiz question's option. The name 'Jacob' has been clarified with the additional detail that Jacob is also referred to as 'Israel'. This change aims to prevent any potential confusion for users who may only be familiar with one name.